Reorganize front end test files
